Move ClusterMode from main.rs into cluster.rs
